feat: update dependencie + endpoint geojson

+++ src/main/java/com/cpe/emergencymanager/model/StationEntity.java
@@ -14,10 +14,10 @@ public class StationEntity {
     private Integer id;
     @Basic
     @Column(name = "longitude")
-    private Float longitude;
+    private Double longitude;
     @Basic
     @Column(name = "latitude")
-    private Float latitude;
+    private Double latitude;
     @OneToMany(mappedBy = "idStation")
     private Collection<FiremanEntity> firemenById;
     @OneToMany(mappedBy = "idStation")
@@ -31,19 +31,19 @@ public class StationEntity {
         this.id = id;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
